+++ src/injectors/dev.ts
@@ -1,5 +1,5 @@
 process.env.NODE_ENV = 'development'
 
-process.on('unhandledRejection', err => {
+process.on('unhandledRejection', (err) => {
   throw err
 })