+++ src/scripts/build.ts
@@ -23,7 +23,7 @@ const build = (): Promise<{ stats: webpack.Stats; warnings: string[] }> => {
       if (err) {
         return reject(err)
       }
-      const messages = formatWebpackMessages(stats.toJson())
+      const messages = formatWebpackMessages(stats!.toJson())
       if (messages.errors.length) {
         if (messages.errors.length > 1 && !RuntimeOptions.noCollapse) {
           messages.errors.length = 1
@@ -50,7 +50,7 @@ const build = (): Promise<{ stats: webpack.Stats; warnings: string[] }> => {
       }
 
       return resolve({
-        stats,
+        stats: stats!,
         warnings: messages.warnings,
       })
     })
@@ -76,7 +76,7 @@ build()
         )
       }
     },
-    err => {
+    (err) => {
       console.log(chalk.red('Failed to compile.\n'))
       console.log(err.message)
       console.log()
@@ -92,7 +92,7 @@ build()
     )
     console.log()
   })
-  .catch(err => {
+  .catch((err) => {
     if (err && err.message) {
       console.log(err.message)
     }