+++ src/scripts/migration/tslint-to-eslint/filesystem.ts
@@ -15,9 +15,9 @@ export function loadJson(path: string): Promise<any> {
   })
 }
 
-export function saveJson(path: string, data: object): Promise<true> {
+export function saveJson(path: string, data: unknown): Promise<true> {
   return new Promise((resolve, reject) => {
-    fs.writeFile(path, JSON.stringify(data, null, 2), err => {
+    fs.writeFile(path, JSON.stringify(data, null, 2), (err) => {
       if (err) {
         return reject(err)
       } else {