+++ src/scripts/migration/tslint-to-eslint/hash.ts
@@ -2,7 +2,7 @@ import crypto from 'crypto'
 import * as fs from 'fs'
 
 export function getFileHash(path: string) {
-  return new Promise(resolve => {
+  return new Promise((resolve) => {
     const hash = crypto.createHash('sha256')
     const input = fs.createReadStream(path, { encoding: 'utf-8' })
     input.on('readable', () => {